Check In
Making sure things haven't become broken since I last looked at the project. Removed the tri material content as it was incompatible with sprites and would be hard to test in this project. Imported some textures to test that method, and fixed the code to actually get a pixel of the texture rather than the player's position in the world, which did not correlate to a position on the texture alone.